# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MMO Battle Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /player:
    get:
      operationId: getPlayer
      tags: [player]
      summary: Get player info
      responses:
        "200":
          description: Player data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: No player yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPlayer
      tags: [player]
      summary: Create or reset a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerRequest"
      responses:
        "200":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /inventory/{id}/equip:
    post:
      operationId: equipItem
      tags: [inventory]
      summary: Equip an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item equipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/{id}/unequip:
    post:
      operationId: unequipItem
      tags: [inventory]
      summary: Unequip an item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item unequipped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EquipResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/{id}/enchant:
    post:
      operationId: enchantItem
      tags: [inventory]
      summary: Enchant an item (+1 stat bonus, costs 1 enchanting stone)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item enchanted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnchantResponse"
        "400":
          description: No stones or item at max enchant level
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CreatePlayerRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        attack:
          type: integer
        defense:
          type: integer
        gold:
          type: integer
        monstersDefeated:
          type: integer
        enchantingStones:
          type: integer
        meleeSkillLevel:
          type: integer
        meleeSkillXp:
          type: integer
        meleeSkillXpToNext:
          type: integer
        defenseSkillLevel:
          type: integer
        defenseSkillXp:
          type: integer
        defenseSkillXpToNext:
          type: integer
        ascensionLevel:
          type: integer
      required:
        - id
        - name
        - level
        - xp
        - xpToNextLevel
        - hp
        - maxHp
        - attack
        - defense
        - gold
        - monstersDefeated
        - enchantingStones
        - meleeSkillLevel
        - meleeSkillXp
        - meleeSkillXpToNext
        - defenseSkillLevel
        - defenseSkillXp
        - defenseSkillXpToNext
        - ascensionLevel
...
    InventoryItem:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        rarity:
          type: string
          enum: [Common, Uncommon, Rare, Epic, Legendary, Mythic, Divine, Abyssal, Transcendent, Cosmic, Eternal]
        emoji:
          type: string
        goldValue:
          type: integer
        type:
          type: string
          enum: [weapon, armor, boots, gloves, amulet, ring]
        statBonus:
          type: integer
        equipped:
          type: boolean
        enchantLevel:
          type: integer
        obtainedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - rarity
        - emoji
        - goldValue
        - type
        - statBonus
        - equipped
        - enchantLevel
        - obtainedAt
    EquipResponse:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/InventoryItem"
        player:
          $ref: "#/components/schemas/Player"
      required:
        - item
        - player
...
    EnchantResponse:
      type: object
      properties:
        item:
          $ref: "#/components/schemas/InventoryItem"
        player:
          $ref: "#/components/schemas/Player"
      required:
        - item
        - player
...